Camera Quirks, Additional Assets
Tweaked the camera so it will adjust properly to any screen resolution, and game objects will be centered appropriately. The game is somewhat unique in that changing the resolution slightly alters gameplay, but in a way that I don't consider detrimental.  The playing field for each user will be viable, yet specific to each device.